+++ src/worker/loan/utils/models.js
@@ -1,7 +1,7 @@
 const LoanMarket = require('../../../models/LoanMarket')
 const Market = require('../../../models/Market')
 const Email = require('../../../models/Email');
-const EthAddress = require('../../../models/EthAddress');
+const AddressEmail = require('../../../models/AddressEmail');
 
 async function getMarketModels (principal, collateral) {
   const loanMarket = await LoanMarket.findOne({ principal, collateral }).exec()
@@ -13,9 +13,9 @@ async function getMarketModels (principal, collateral) {
   return { loanMarket, market }
 }
 
-async function getEmails(ethAddress) {
-  console.log("Finding emails for ", ethAddress)
-  const res = await EthAddress.findOne({address: ethAddress}).populate({path: 'emails', model: 'Email'}).exec()
+async function getEmails(addressEmail) {
+  console.log("Finding emails for ", addressEmail)
+  const res = await AddressEmail.findOne({address: addressEmail}).populate({path: 'emails', model: 'Email'}).exec()
   console.log(res)
   return res ? res.emails : {}
 }
